Added working periodic, fcc, measuring

+++ atom.cpp
@@ -2,7 +2,7 @@
 #include <atom.h>
 #include <math/random.h>
 
-Atom::Atom(double mass) :
+Atom::Atom(float mass) :
     m_mass(mass)
 {
     
@@ -18,10 +18,10 @@ void Atom::resetForce()
     force.setToZero();
 }
 
-void Atom::resetVelocityMaxwellian(double temperature)
+void Atom::resetVelocityMaxwellian(float temperature)
 {
     // Resetting the velocity according to a Maxwell-Boltzmann distribution (see http://en.wikipedia.org/wiki/Maxwell%E2%80%93Boltzmann_distribution )
-    double boltzmannConstant = 1.0; // In atomic units, the boltzmann constant equals 1
-    double standardDeviation = sqrt(boltzmannConstant*temperature/m_mass);
+    float boltzmannConstant = 1.0; // In atomic units, the boltzmann constant equals 1
+    float standardDeviation = sqrt(boltzmannConstant*temperature/m_mass);
     velocity.randomGaussian(0, standardDeviation);
 }
