+++ atom.h
@@ -5,17 +5,17 @@ using CompPhys::vec3;
 class Atom
 {
 private:
-    double m_mass;
+    float m_mass;
 public:
     vec3 position;
     vec3 velocity;
     vec3 force;
 
-    Atom(double mass);
+    Atom(float mass);
     ~Atom();
     void resetForce();
-    void resetVelocityMaxwellian(double temperature);
+    void resetVelocityMaxwellian(float temperature);
 
-    inline double mass() { return m_mass; }
-    inline void setMass(double mass) { m_mass = mass; }
+    inline float mass() { return m_mass; }
+    inline void setMass(float mass) { m_mass = mass; }
 };
