+++ unitconverter.cpp
@@ -6,22 +6,22 @@
 using namespace std;
 using CompPhys::vec3;
 
-double UnitConverter::m0 = 0;
-double UnitConverter::q0 = 0;
-double UnitConverter::hbar0 = 0;
-double UnitConverter::electricConstant0 = 0;
-double UnitConverter::a0 = 0;
-double UnitConverter::a0_angstrom = 0;
-double UnitConverter::E0 = 0;
-double UnitConverter::E0ev = 0;
-double UnitConverter::kb = 0;
-double UnitConverter::t0 = 0;
-double UnitConverter::F0 = 0;
-double UnitConverter::T0 = 0;
-double UnitConverter::P0 = 0;
-double UnitConverter::v0 = 0;
-double UnitConverter::visc0 = 0;
-double UnitConverter::diff0 = 0;
+float UnitConverter::m0 = 0;
+float UnitConverter::q0 = 0;
+float UnitConverter::hbar0 = 0;
+float UnitConverter::electricConstant0 = 0;
+float UnitConverter::a0 = 0;
+float UnitConverter::a0_angstrom = 0;
+float UnitConverter::E0 = 0;
+float UnitConverter::E0ev = 0;
+float UnitConverter::kb = 0;
+float UnitConverter::t0 = 0;
+float UnitConverter::F0 = 0;
+float UnitConverter::T0 = 0;
+float UnitConverter::P0 = 0;
+float UnitConverter::v0 = 0;
+float UnitConverter::visc0 = 0;
+float UnitConverter::diff0 = 0;
 std::string UnitConverter::currentUnits = "No units chosen.";
 
 bool UnitConverter::initialized = false;
@@ -32,11 +32,11 @@ void UnitConverter::initializeMDUnits() {
     // Molecular Dynamics units
 
     // Fundamental units
-    double m0 = 1.66053892e-27;         // SI [kg]
-    double L0 = 1e-10;                  // SI [m]
-    double kb = 1.3806488e-23;          // SI [J/K]
-    double E0eV = 1.0318e-2;            // eV
-    double E0 = 1.60217657e-19*E0eV;    // SI [J]
+    float m0 = 1.66053892e-27;         // SI [kg]
+    float L0 = 1e-10;                  // SI [m]
+    float kb = 1.3806488e-23;          // SI [J/K]
+    float E0eV = 1.0318e-2;            // eV
+    float E0 = 1.60217657e-19*E0eV;    // SI [J]
 
     UnitConverter::m0 = m0;
     UnitConverter::kb = kb;
@@ -64,11 +64,11 @@ void UnitConverter::initializeAtomicUnits() {
     // [1] http://en.wikipedia.org/wiki/Atomic_units
 
     // Fundamental units
-    double m0 = 9.10938291e-31;  // SI [kg]
-    double q0 = 1.602176565e-19; // SI [C]
-    double hbar0 = 1.054571726e-34; // SI [Js]
-    double electricConstant0 = 8.9875517873681e9; // SI [kgm^3/(s^-2 C^-2)]
-    double kb = 1.3806488e-23; // SI [J/K]
+    float m0 = 9.10938291e-31;  // SI [kg]
+    float q0 = 1.602176565e-19; // SI [C]
+    float hbar0 = 1.054571726e-34; // SI [Js]
+    float electricConstant0 = 8.9875517873681e9; // SI [kgm^3/(s^-2 C^-2)]
+    float kb = 1.3806488e-23; // SI [J/K]
 
     UnitConverter::m0 = m0;
     UnitConverter::q0 = q0;
@@ -99,20 +99,20 @@ void UnitConverter::initialize(Units type) {
     else if(type == MDUnits) UnitConverter::initializeMDUnits();
 }
 
-double UnitConverter::pressureToSI(double P) {UnitConverter::makeSureInitialized(); return UnitConverter::P0*P; }
-double UnitConverter::pressureFromSI(double P) {UnitConverter::makeSureInitialized(); return P/UnitConverter::P0; }
+float UnitConverter::pressureToSI(float P) {UnitConverter::makeSureInitialized(); return UnitConverter::P0*P; }
+float UnitConverter::pressureFromSI(float P) {UnitConverter::makeSureInitialized(); return P/UnitConverter::P0; }
 
-double UnitConverter::temperatureToSI(double T) {UnitConverter::makeSureInitialized(); return UnitConverter::T0*T; }
-double UnitConverter::temperatureFromSI(double T) {UnitConverter::makeSureInitialized(); return T/UnitConverter::T0; }
+float UnitConverter::temperatureToSI(float T) {UnitConverter::makeSureInitialized(); return UnitConverter::T0*T; }
+float UnitConverter::temperatureFromSI(float T) {UnitConverter::makeSureInitialized(); return T/UnitConverter::T0; }
 
-double UnitConverter::massToSI(double m) {UnitConverter::makeSureInitialized(); return UnitConverter::m0*m; }
-double UnitConverter::massFromSI(double m) {UnitConverter::makeSureInitialized(); return m/UnitConverter::m0; }
+float UnitConverter::massToSI(float m) {UnitConverter::makeSureInitialized(); return UnitConverter::m0*m; }
+float UnitConverter::massFromSI(float m) {UnitConverter::makeSureInitialized(); return m/UnitConverter::m0; }
 
-double UnitConverter::lengthToSI(double L) {UnitConverter::makeSureInitialized(); return UnitConverter::a0*L; }
-double UnitConverter::lengthFromSI(double L) {UnitConverter::makeSureInitialized(); return L/UnitConverter::a0; }
+float UnitConverter::lengthToSI(float L) {UnitConverter::makeSureInitialized(); return UnitConverter::a0*L; }
+float UnitConverter::lengthFromSI(float L) {UnitConverter::makeSureInitialized(); return L/UnitConverter::a0; }
 
-double UnitConverter::lengthToAngstroms(double L) {UnitConverter::makeSureInitialized(); return UnitConverter::a0*L*1e10; }
-double UnitConverter::lengthFromAngstroms(double L) {UnitConverter::makeSureInitialized(); return L/(UnitConverter::a0*1e10); }
+float UnitConverter::lengthToAngstroms(float L) {UnitConverter::makeSureInitialized(); return UnitConverter::a0*L*1e10; }
+float UnitConverter::lengthFromAngstroms(float L) {UnitConverter::makeSureInitialized(); return L/(UnitConverter::a0*1e10); }
 
 vec3 UnitConverter::lengthToSI(vec3 position)
 {
@@ -144,23 +144,23 @@ CompPhys::vec3 UnitConverter::velocityFromSI(CompPhys::vec3 velocity)
     return vec3(UnitConverter::velocityFromSI(velocity.x()), UnitConverter::velocityFromSI(velocity.y()), UnitConverter::velocityFromSI(velocity.z()));
 }
 
-double UnitConverter::forceToSI(double F) {UnitConverter::makeSureInitialized(); return UnitConverter::F0*F; }
-double UnitConverter::forceFromSI(double F) {UnitConverter::makeSureInitialized(); return F/UnitConverter::F0; }
+float UnitConverter::forceToSI(float F) {UnitConverter::makeSureInitialized(); return UnitConverter::F0*F; }
+float UnitConverter::forceFromSI(float F) {UnitConverter::makeSureInitialized(); return F/UnitConverter::F0; }
 
-double UnitConverter::energyToSI(double E) {UnitConverter::makeSureInitialized(); return UnitConverter::E0*E; }
-double UnitConverter::energyFromSI(double E) {UnitConverter::makeSureInitialized(); return E/UnitConverter::E0; }
+float UnitConverter::energyToSI(float E) {UnitConverter::makeSureInitialized(); return UnitConverter::E0*E; }
+float UnitConverter::energyFromSI(float E) {UnitConverter::makeSureInitialized(); return E/UnitConverter::E0; }
 
-double UnitConverter::energyToEv(double E) {UnitConverter::makeSureInitialized(); return UnitConverter::E0ev*E; }
-double UnitConverter::energyFromEv(double E) {UnitConverter::makeSureInitialized(); return E/UnitConverter::E0ev; }
+float UnitConverter::energyToEv(float E) {UnitConverter::makeSureInitialized(); return UnitConverter::E0ev*E; }
+float UnitConverter::energyFromEv(float E) {UnitConverter::makeSureInitialized(); return E/UnitConverter::E0ev; }
 
-double UnitConverter::degreesToRadians(double v) {UnitConverter::makeSureInitialized(); return M_PI/180*v; }
-double UnitConverter::radiansToDegrees(double v) {UnitConverter::makeSureInitialized(); return 180/M_PI*v; }
+float UnitConverter::degreesToRadians(float v) {UnitConverter::makeSureInitialized(); return M_PI/180*v; }
+float UnitConverter::radiansToDegrees(float v) {UnitConverter::makeSureInitialized(); return 180/M_PI*v; }
 
-double UnitConverter::timeToSI(double t) {UnitConverter::makeSureInitialized(); return UnitConverter::t0*t; }
-double UnitConverter::timeFromSI(double t) {UnitConverter::makeSureInitialized(); return t/UnitConverter::t0; }
+float UnitConverter::timeToSI(float t) {UnitConverter::makeSureInitialized(); return UnitConverter::t0*t; }
+float UnitConverter::timeFromSI(float t) {UnitConverter::makeSureInitialized(); return t/UnitConverter::t0; }
 
-double UnitConverter::velocityToSI(double v) {UnitConverter::makeSureInitialized(); return v*UnitConverter::v0; }
-double UnitConverter::velocityFromSI(double v) {UnitConverter::makeSureInitialized(); return v/UnitConverter::v0; }
+float UnitConverter::velocityToSI(float v) {UnitConverter::makeSureInitialized(); return v*UnitConverter::v0; }
+float UnitConverter::velocityFromSI(float v) {UnitConverter::makeSureInitialized(); return v/UnitConverter::v0; }
 
-double UnitConverter::diffusionToSI(double d) {UnitConverter::makeSureInitialized(); return d*UnitConverter::diff0; }
-double UnitConverter::diffusionFromSI(double d) {UnitConverter::makeSureInitialized(); return d/UnitConverter::diff0; }
+float UnitConverter::diffusionToSI(float d) {UnitConverter::makeSureInitialized(); return d*UnitConverter::diff0; }
+float UnitConverter::diffusionFromSI(float d) {UnitConverter::makeSureInitialized(); return d/UnitConverter::diff0; }
